Store the timeout as an Integer
The timeout expects an integer, but currently it is a Long.

+++ src/main/groovy/net/ll86/plugins/VpnPluginExtension.groovy
@@ -2,6 +2,6 @@ package net.ll86.plugins
 
 class VpnPluginExtension {
     String baseHost 
-    Long pingTimeout
+    Integer pingTimeout
 }
